+++ util/lib/Style.mli
@@ -28,7 +28,14 @@ val left_col : string -> string
 (*
    Print multiline text with an indentation. Always adds a trailing newline.
 
+   The midsection is elided if the input string exceeds 'max_bytes'.
+   Data fragments can be highlighted in a different color than the
+   text we insert to signal the elision.
+
    Pros: disambiguates the quoted text from other output
    Cons: interferes with proper copy-pasting (since it inserts whitespace)
 *)
-val quote_multiline_text : string -> string
+val quote_multiline_text :
+  ?decorate_comment:(string -> string) ->
+  ?decorate_data_fragment:(string -> string) ->
+  ?max_bytes:int -> string -> string